+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Build map list dynamically
 - Add map list for multiplanetary systems with only one starting planet
+- Changed mod name from Galactic War Multiplanetary Systems to Multiplanetary System Tabs
 
 ## v1.0.3 - 2021-12-04